+++ test/scripts/lint/file.js
@@ -1,0 +1,6 @@
+i18n.t("js.missing");
+i18n.t("js.missing", { scope: "base" });
+i18n.t("bunny rabbit adventure");
+i18n.t("js.missing", { defaults: [{ scope: "bunny rabbit adventure" }] });
+i18n.t("ignore_scope");
+i18n.t("another_ignore_scope");